Rewrite in CoffeeScript, refactor model/UI relationships in a few places
Breaks down a few things, mainly related to CS for..in list comprehension
loops, as compared to JS for..in object property enumeration loops. This
should be fixed by using a new `Hashtable` class where of dumb {} objects
were used instead.

`PlayerGL` is a new word for `PlayerObject`, and `NodeVsZombiesUI` was
separated into `World` and `WorldGL`. `PlayerGL` extends `Player` and
expects to belong to `WorldGL`.